integrated convertPosIntToWords into PigNumberClass class
